+++ src/aks-preview/HISTORY.rst
@@ -11,7 +11,18 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+
+4.0.0b3
++++++++
+* Leave only one role assignment for automatic sku clusters.
+    * "Azure Kubernetes Service RBAC Cluster Admin"
+
+4.0.0b2
++++++++
 * Improve Windows OutboundNat test case by removing Windows OSSKU limitation
+* `az aks create/update`: add support for new outbound type none
+* Add `az operation show` command to show the details of a specific operation.
+* Add `az operation show-latest` command to show the details of the latest operation.
 
 4.0.0b1
 +++++++